// web/src/lib/types/user.ts
type user = {
  accent_color?: number;
  avatar?: string;
  avatar_decoration_data?: any;
  avatar_url?: string;
  banner?: any;
  banner_color?: string;
  discriminator?: string;
  email?: string;
  flags?: number;
  global_name?: string;
  id?: string;
  locale?: string;
  mfa_enabled?: boolean;
  name?: string;
  premium_type?: number;
  public_flags?: number;
  username?: string;
  verified?: boolean;
};

export default user;